# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Aquachile",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/productivity": {
            "get": {
                "tags": [
                    "KPI Productivity"
                ],
                "summary": "Show all productivity kpi",
                "description": "Returns a list of productivity kpi",
                "operationId": "38744025d7cc0ac4204295b71f0cd846",
                "responses": {
                    "200": {
                        "description": "Show all productivity kpi."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/productivity-line/{id}": {
            "get": {
                "tags": [
                    "KPI Productivity"
                ],
                "summary": "Show all productivity kpi by line",
                "description": "Returns a list of productivity kpi by line",
                "operationId": "5cc07ecaa0d6bfe4538fe12f68a36f72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Line ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all productivity kpi by line."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/productivity-shift/{id}": {
            "get": {
                "tags": [
                    "KPI Productivity"
                ],
                "summary": "Show all productivity kpi by shift",
                "description": "Returns a list of productivity kpi by shift",
                "operationId": "718c7244d6504eaa1ef65a6c129565ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Shift ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all productivity kpi by shift."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/productivity/{id}": {
            "post": {
                "tags": [
                    "KPI Productivity"
                ],
                "summary": "Create a new productivity KPI",
                "description": "Creates a new productivity KPI with the received data",
                "operationId": "0bde265abd6d6c5ef94e9cb9667dd296",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Line ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Productivity KPI  object that needs to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "packed_tons",
                                    "tons_produced",
                                    "line_name"
                                ],
                                "properties": {
                                    "packed_tons": {
                                        "description": "The packed tons of the KPI",
                                        "type": "string"
                                    },
                                    "tons_produced": {
                                        "description": "The tons produced of the KPI",
                                        "type": "string"
                                    },
                                    "uuid": {
                                        "description": "The shift uuid",
                                        "type": "uuid"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "packed_tons": 4,
                                    "tons_produced": 1000,
                                    "uuid": "d7f3c2f3-0d4c-4f2e-8b5e-4e6c3c3f4e2c"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}